rewriting the lexxer
