+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod lex;
-//mod parser;
+mod lex_2;
+
+
 //mod ast;
 
 pub use lex::*;
