+++ src/token_type.rs
@@ -1,0 +1,23 @@
+
+
+/// PYRUST TOKEN TYPE
+
+#[derive(Debug, PartialEq,Eq, Clone)]
+
+pub enum TokenType{
+    ENDMARKER,
+    EOF,
+    NAME,
+    NUMBER,
+    STRING,
+    NEWLINE,
+    INDENT,
+    DEDENT,
+
+
+
+    ///OPERATORS
+
+
+
+}
